Dev -> Prod

+++ modules/storage/gcp_storage.go
@@ -83,6 +83,8 @@ func (g *GCPStorage) CopyFromBytesToRemote(inputBytes []byte, instanceNames []st
 		return err
 	}
 
+	// If we error somewhere in the loop, don't exit until all instances have been tried (debug instance isn't live)
+	var loopError error
 	for _, name := range instanceNames {
 		// Call gsutil to copy the tmp file over to the instance
 		runnable := exec.Command("gcloud", "compute", "scp", "--zone", "us-central1-a", outputFileName, fmt.Sprintf("%s:%s", name, outputFileName))
@@ -91,10 +93,13 @@ func (g *GCPStorage) CopyFromBytesToRemote(inputBytes []byte, instanceNames []st
 		level.Debug(g.Logger).Log("msg", buffer)
 		if err != nil {
 			err = fmt.Errorf("failed to copy file to instance: %v", err)
-			return err
+			loopError = err
 		}
 	}
 
+	if loopError != nil {
+		return loopError
+	}
 	// delete temp file to clean up
 	if err := os.Remove(outputFileName); err != nil {
 		err = fmt.Errorf("failed to clean up tmp file: %v", err)
@@ -114,42 +119,7 @@ func (g *GCPStorage) CopyFromBucketToLocal(ctx context.Context, artifactName str
 		}
 	}
 
-	// setup reader
-	reader, err := g.Bucket.Object(artifactName).NewReader(ctx)
-	if err != nil {
-		err = fmt.Errorf("failed to create bucket reader: %v", err)
-		return err
-	}
-
-	defer reader.Close()
-
-	// read in artifact as bytes to be copied to new file
-	var artifactBytes []byte
-	_, err = reader.Read(artifactBytes)
-	if err != nil {
-		err = fmt.Errorf("failed to read artifact: %v", err)
-		return err
-	}
-
-	if err := ioutil.WriteFile(outputLocation, artifactBytes, 0644); err != nil {
-		err = fmt.Errorf("failed to write local file: %v", err)
-		return err
-	}
-
-	return nil
-}
-
-// Copy artifact to remote file location (SCP function)
-func (g *GCPStorage) CopyFromBucketToRemote(ctx context.Context, artifactName string, srcFileName string, instanceName string, outputFileName string) error {
-	// grab file from bucket
-	if err := g.CopyFromBucketToLocal(ctx, artifactName, srcFileName); err != nil {
-		err = fmt.Errorf("failed to fetch file from GCP bucket: %v", err)
-		return err
-	}
-
-	// Call gsutil to copy the tmp file over to the instance
-	runnable := exec.Command("gcloud", "compute", "scp", "--zone", "us-central1-a", outputFileName, fmt.Sprintf("%s:%s", instanceName, outputFileName))
-
+	runnable := exec.Command("gsutil", "cp", artifactName, outputLocation)
 	buffer, err := runnable.CombinedOutput()
 
 	level.Debug(g.Logger).Log("msg", buffer)
@@ -158,8 +128,38 @@ func (g *GCPStorage) CopyFromBucketToRemote(ctx context.Context, artifactName st
 		return err
 	}
 
+	return nil
+}
+
+// Copy artifact to remote file location (SCP function)
+func (g *GCPStorage) CopyFromBucketToRemote(ctx context.Context, artifactName string, instanceNames []string, outputFileName string) error {
+	// grab file from bucket
+	if err := g.CopyFromBucketToLocal(ctx, artifactName, outputFileName); err != nil {
+		err = fmt.Errorf("failed to fetch file from GCP bucket: %v", err)
+		return err
+	}
+
+	// If we error somewhere in the loop, don't exit until all instances have been tried (debug instance isn't live)
+	var loopError error
+	for _, name := range instanceNames {
+		// Call gsutil to copy the tmp file over to the instance
+		runnable := exec.Command("gcloud", "compute", "scp", "--zone", "us-central1-a", outputFileName, fmt.Sprintf("%s:%s", name, outputFileName))
+
+		buffer, err := runnable.CombinedOutput()
+
+		level.Debug(g.Logger).Log("msg", buffer)
+		if err != nil {
+			err = fmt.Errorf("failed to copy file to instance: %v", err)
+			loopError = err
+		}
+	}
+
+	if loopError != nil {
+		return loopError
+	}
+
 	// delete src file to clean up
-	if err := os.Remove(srcFileName); err != nil {
+	if err := os.Remove(outputFileName); err != nil {
 		err = fmt.Errorf("failed to clean up tmp file: %v", err)
 		return err
 	}
